chore(autopipeline): consolidate pipeline test files; mark the feature experimental
Consolidate the five loose pipeline test files into two, mirroring the
engine PR's test consolidation: the dedicated-pool battery (handshake,
maintnotifications, streaming credentials) into pipeline_pool_test.go and
the execution-behavior tests (heterogeneous reply-shape demux, the
retry/re-drive loop) into pipeline_exec_test.go. Pure move — ten test
functions before and after, with section markers keeping provenance.

Mark autopipelining EXPERIMENTAL using the repo's existing marker line:
the API may still change as feedback lands. The marker covers every public
surface — the eight client getters (Client and ClusterClient, blocking and
async, no-arg and WithOptions), the AutoPipeliner and AutoPipelineOptions
types, both Default*Options constructors, the UniversalClient interface
block and Options.AutoPipelineOptions — plus banners in the README
autopipelining section and the example.
